Extract manifest->bundle.json logic into its own pkg
