+++ pkg/cnab/config_adapter/adapter_test.go
@@ -1,0 +1,81 @@
+package configadapter
+
+import (
+	"testing"
+
+	"github.com/deislabs/porter/pkg/config"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+func TestPorter_generateImages(t *testing.T) {
+	c := config.NewTestConfig(t)
+	c.TestContext.AddTestFile("../../config/testdata/simple.porter.yaml", config.Name)
+
+	err := c.LoadManifest()
+	require.NoError(t, err)
+
+	a := ManifestConverter{
+		Context:  c.Context,
+		Manifest: c.Manifest,
+	}
+
+	mappedImage := config.MappedImage{
+		Description:   "un petite server",
+		Image:         "deislabs/myserver:1.0.0",
+		ImageType:     "docker",
+		Digest:        "abc123",
+		Size:          12,
+		MediaType:     "download",
+		OriginalImage: "deis/myserver:1.0.0",
+		Platform: &config.ImagePlatform{
+			OS:           "linux",
+			Architecture: "amd64",
+		},
+	}
+	a.Manifest.ImageMap = map[string]config.MappedImage{
+		"server": mappedImage,
+	}
+
+	images := a.generateBundleImages()
+
+	require.Len(t, images, 1)
+	img := images["server"]
+	assert.Equal(t, mappedImage.Description, img.Description)
+	assert.Equal(t, mappedImage.Image, img.Image)
+	assert.Equal(t, mappedImage.ImageType, img.ImageType)
+	assert.Equal(t, mappedImage.Digest, img.Digest)
+	assert.Equal(t, mappedImage.Size, img.Size)
+	assert.Equal(t, mappedImage.MediaType, img.MediaType)
+	assert.Equal(t, mappedImage.OriginalImage, img.OriginalImage)
+	assert.Equal(t, mappedImage.Platform.OS, img.Platform.OS)
+	assert.Equal(t, mappedImage.Platform.Architecture, img.Platform.Architecture)
+}
+
+func TestPorter_generateBundleImages_EmptyPlatform(t *testing.T) {
+	c := config.NewTestConfig(t)
+	c.TestContext.AddTestFile("../../config/testdata/simple.porter.yaml", config.Name)
+
+	err := c.LoadManifest()
+	require.NoError(t, err)
+
+	a := ManifestConverter{
+		Context:  c.Context,
+		Manifest: c.Manifest,
+	}
+
+	mappedImage := config.MappedImage{
+		Description: "un petite server",
+		Image:       "deislabs/myserver:1.0.0",
+		ImageType:   "docker",
+		Platform:    nil,
+	}
+	a.Manifest.ImageMap = map[string]config.MappedImage{
+		"server": mappedImage,
+	}
+
+	images := a.generateBundleImages()
+	require.Len(t, images, 1)
+	img := images["server"]
+	assert.Nil(t, img.Platform)
+}
