+++ pkg/cnab/config_adapter/stamp.go
@@ -1,4 +1,4 @@
-package config
+package configadapter
 
 import (
 	"crypto/sha256"
@@ -7,6 +7,7 @@ import (
 	"fmt"
 
 	"github.com/deislabs/cnab-go/bundle"
+	"github.com/deislabs/porter/pkg/config"
 	"github.com/pkg/errors"
 )
 
@@ -14,10 +15,10 @@ type Stamp struct {
 	ManifestDigest string `json:"manifestDigest"`
 }
 
-func (c *Config) GenerateStamp(m *Manifest) Stamp {
+func (c *ManifestConverter) GenerateStamp() Stamp {
 	stamp := Stamp{}
 
-	digest, err := c.digestManifest(m)
+	digest, err := c.digestManifest()
 	if err != nil {
 		// The digest is only used to decide if we need to rebuild, it is not an error condition to not
 		// have a digest.
@@ -30,22 +31,22 @@ func (c *Config) GenerateStamp(m *Manifest) Stamp {
 	return stamp
 }
 
-func (c *Config) digestManifest(m *Manifest) (string, error) {
-	if exists, _ := c.FileSystem.Exists(m.path); !exists {
-		return "", errors.Errorf("the specified porter configuration file %s does not exist", m.path)
+func (c *ManifestConverter) digestManifest() (string, error) {
+	if exists, _ := c.FileSystem.Exists(c.Manifest.GetManifestPath()); !exists {
+		return "", errors.Errorf("the specified porter configuration file %s does not exist", c.Manifest.GetManifestPath())
 	}
 
-	data, err := c.FileSystem.ReadFile(m.path)
+	data, err := c.FileSystem.ReadFile(c.Manifest.GetManifestPath())
 	if err != nil {
-		return "", errors.Wrapf(err, "could not read manifest at %q", m.path)
+		return "", errors.Wrapf(err, "could not read manifest at %q", c.Manifest.GetManifestPath())
 	}
 
 	digest := sha256.Sum256(data)
 	return hex.EncodeToString(digest[:]), nil
 }
 
-func (c *Config) LoadStamp(bun *bundle.Bundle) (*Stamp, error) {
-	data := bun.Custom[CustomBundleKey]
+func LoadStamp(bun *bundle.Bundle) (*Stamp, error) {
+	data := bun.Custom[config.CustomBundleKey]
 
 	dataB, err := json.Marshal(data)
 	if err != nil {
