+++ pkg/porter/build_test.go
@@ -5,6 +5,7 @@ import (
 	"testing"
 
 	"github.com/deislabs/cnab-go/bundle"
+	configadapter "github.com/deislabs/porter/pkg/cnab/config_adapter"
 	"github.com/deislabs/porter/pkg/config"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -196,7 +197,7 @@ func TestPorter_buildBundle(t *testing.T) {
 	assert.Equal(t, bun.Version, "0.1.0")
 	assert.Equal(t, bun.Description, "An example Porter configuration")
 
-	stamp, err := p.LoadStamp(bun)
+	stamp, err := configadapter.LoadStamp(bun)
 	require.NoError(t, err)
 	assert.Equal(t, "06a51d04297375bf111ab15e579b8a7ab72e2661018c4d08d1d3f38198028e49", stamp.ManifestDigest)
 
@@ -232,74 +233,4 @@ func TestPorter_paramRequired(t *testing.T) {
 	p2, ok := bundle.Parameters["command2"]
 	require.True(t, ok)
 	require.True(t, p2.Required)
-}
-
-func TestPorter_generateImages(t *testing.T) {
-	p := NewTestPorter(t)
-	p.TestConfig.SetupPorterHome()
-
-	configTpl, err := p.Templates.GetManifest()
-	require.Nil(t, err)
-	p.TestConfig.TestContext.AddTestFileContents(configTpl, config.Name)
-
-	err = p.LoadManifest()
-	require.NoError(t, err)
-
-	mappedImage := config.MappedImage{
-		Description:   "un petite server",
-		Image:         "deislabs/myserver:1.0.0",
-		ImageType:     "docker",
-		Digest:        "abc123",
-		Size:          12,
-		MediaType:     "download",
-		OriginalImage: "deis/myserver:1.0.0",
-		Platform: &config.ImagePlatform{
-			OS:           "linux",
-			Architecture: "amd64",
-		},
-	}
-	p.Manifest.ImageMap = map[string]config.MappedImage{
-		"server": mappedImage,
-	}
-
-	images := p.generateBundleImages()
-
-	require.Len(t, images, 1)
-	img := images["server"]
-	assert.Equal(t, mappedImage.Description, img.Description)
-	assert.Equal(t, mappedImage.Image, img.Image)
-	assert.Equal(t, mappedImage.ImageType, img.ImageType)
-	assert.Equal(t, mappedImage.Digest, img.Digest)
-	assert.Equal(t, mappedImage.Size, img.Size)
-	assert.Equal(t, mappedImage.MediaType, img.MediaType)
-	assert.Equal(t, mappedImage.OriginalImage, img.OriginalImage)
-	assert.Equal(t, mappedImage.Platform.OS, img.Platform.OS)
-	assert.Equal(t, mappedImage.Platform.Architecture, img.Platform.Architecture)
-}
-
-func TestPorter_generateBundleImages_EmptyPlatform(t *testing.T) {
-	p := NewTestPorter(t)
-	p.TestConfig.SetupPorterHome()
-
-	configTpl, err := p.Templates.GetManifest()
-	require.Nil(t, err)
-	p.TestConfig.TestContext.AddTestFileContents(configTpl, config.Name)
-
-	err = p.LoadManifest()
-	require.NoError(t, err)
-
-	mappedImage := config.MappedImage{
-		Description: "un petite server",
-		Image:       "deislabs/myserver:1.0.0",
-		ImageType:   "docker",
-		Platform:    nil,
-	}
-	p.Manifest.ImageMap = map[string]config.MappedImage{
-		"server": mappedImage,
-	}
-
-	images := p.generateBundleImages()
-	require.Len(t, images, 1)
-	img := images["server"]
-	assert.Nil(t, img.Platform)
 }
