+++ pkg/porter/stamp.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 
 	"github.com/deislabs/cnab-go/bundle"
+	configadapter "github.com/deislabs/porter/pkg/cnab/config_adapter"
 	"github.com/pkg/errors"
 )
 
@@ -39,12 +40,16 @@ func (p *Porter) IsBundleUpToDate(opts bundleFileOptions) (bool, error) {
 			return false, errors.Wrapf(err, "could not marshal data from %s", opts.CNABFile)
 		}
 
-		oldStamp, err := p.LoadStamp(bun)
+		oldStamp, err := configadapter.LoadStamp(bun)
 		if err != nil {
 			return false, errors.Wrapf(err, "could not load stamp from %s", opts.CNABFile)
 		}
 
-		newStamp := p.GenerateStamp(p.Manifest)
+		converter := configadapter.ManifestConverter{
+			Context:  p.Context,
+			Manifest: p.Manifest,
+		}
+		newStamp := converter.GenerateStamp()
 		return oldStamp.ManifestDigest == newStamp.ManifestDigest, nil
 	}
